TSP-387 Fixed Incorrect type of referrer id (#85)

+++ src/main/java/com/github/sportbot/repository/UserRepository.java
@@ -23,6 +23,6 @@ public interface UserRepository extends JpaRepository<User, Integer> {
         FROM User u
         WHERE u.referrerTelegramId = :referrerTelegramId
     """)
-    Integer countByReferrerTelegramId(@Param("referrerTelegramId") Integer referrerTelegramId);
+    Integer countByReferrerTelegramId(@Param("referrerTelegramId") Long referrerTelegramId);
 }
